utils/gcp_docker_sft/src/requirements.txt
packaging
flash-attn
wandb
transformers
datasets
accelerate
peft
trl
bitsandbytes
google-cloud-secret-manager
huggingface-hub
python-dotenv
